grading-engine/requirements.txt
celery==5.3.4
redis==5.0.1
requests==2.31.0
openai==1.6.0
PyYAML==6.0.1
chromadb>=1.5.0
sentence-transformers>=5.0
docker==7.0.0
